# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Monopoly Empire API
  version: 1.0.0
servers:

  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/west-fd6/empire/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/uvt-636/empire/1.0.0
  - url: http://localhost:8080
    description: Local server
...
paths:
  /api/jocuri:
    get:
      tags:
        - joc
      summary: Returnam toate jocurile
      description: Returnam toate jocurile
      operationId: getGames
      responses:
        '200':
          description: Jocurile au fost returnate cu succes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Joc'          
    post:
      tags:
        - joc
      summary: Adaugarea unui joc nou
      description: Crearea unui joc prin specificarea numarului de jucatori. Se va genera un id al jocului si se va adauga jucatorul care a creat jocul in lista cu jucatori, de asemnea statusul jocului va fi "WATING"
      operationId: creareJoc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Joc-creareJoc'
        required: true
      responses:
        '200':
          description: Jocul a fost creat cu succes
...
  /api/jocuri/{idJoc}:
    get:
      tags:
        - joc
      summary: Gaseste jocul in functie de id-ul acestia
      description: Returnam datele despre jocul corespunzator id-ulu specificat
      operationId: findGameById
      parameters:
        - name: idJoc
          in: path
          description: id-ul jocului (numar intreg pozitiv)
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Jocul a fost returnat cu succes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Joc'          
        '400':
          description: Invalid id
      security:
        - petstore_auth:
            - write:pets
            - read:pets
  /api/jocuri/alaturareJoc/{idJoc}:
    post:
      tags:
        - joc
      parameters:
        - name: idJoc
          in: path
          required: true
          schema:
            type: integer
      summary: Adaugarea unui jucator la un joc
      description: Adaugarea unui jucator la un joc folosind username-ul jucatorului si id-ul jocului
      operationId: actualizareJoc
      responses:
        '200':
          description: Jucator adaugat cu succes
        '404':
          description: Id-ul jocului nu a fost gasit.Username-ul jucatorului nu a fost gasit.
  /api/jocuri/{idJoc}/jucatori:
    get:
      tags:
        - joc
      summary: Retruneaza toti jucatorii unui joc
      description: Retruneaza toti jucatorii unui joc
      operationId: getGamePlayers
      parameters:
        - name: idJoc
          in: path
          description: id-ul jocului (numar intreg pozitiv)
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Jucatorii au fost returnati cu succes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Joc-returnareJucatoriJoc'          
        '400':
          description: Invalid id
  /api/jocuri/{idJoc}/turnuri:
    get:
      tags:
        - joc
      summary: Afisarea detaliilor despre turnul jucatorilor din jocul cu id-ul specificat
      operationId: listTurnJoc
      parameters:
        - name: idJoc
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Descriere turn
          content:
            application/json:
              schema:   
                $ref: '#/components/schemas/Turn'
        '404':
          description: Jocul nu a fost gasit
  /api/jocuri/{idJoc}/carduri:
    get:
      tags:
        - joc
      summary: Extragerea informatiilor despre toate cardurile pe care le detin toti jucatorii din jocul cu id-ul specificat
      description: Extragerea informatiilor despre toate cardurile pe care le detin toti jucatorii din jocul cu id-ul specificat
      operationId: getCardsByGameId
      parameters:
        - name: idJoc
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalii despre carduri returnate cu succes
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CarduriActiveByUsername'
        '404':
          description: Username-ul jucatorului nu a fost gasit.
  /api/jocuri/{idJoc}/panouri:
    get:
      tags:
        - joc
      summary: Returneaza toate panourile cumparate intr-un joc
      description: Returneaza toate panourile cumparate intr-un joc
      operationId: listPanouCumparat
      parameters:
        - name: idJoc
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista panourilor cumparate
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Panou'
        '404':
          description: Jocul nu a fost gasit
...
  /api/jocuri/parasireJoc/{idJoc}:
    post:
      tags:
        - joc
      summary: Părăsirea unui joc de către un jucător
      description: Permite unui jucător să iasă dintr-un joc, eliminându-l din lista de jucători
      operationId: parasireJoc
      parameters:
        - name: idJoc
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Jucătorul a părăsit jocul cu succes
        '400':
          description: Utilizatorul nu este în acest joc
        '404':
          description: Jocul nu a fost găsit

  # Rută pentru verificarea jocului curent al unui utilizator
  /api/jocuri/jocCurent:
    get:
      tags:
        - joc
      summary: Verifică jocul curent al utilizatorului
      description: Returnează informații despre jocul în care se află utilizatorul curent
      operationId: getJocCurent
      responses:
        '200':
          description: Informații despre jocul curent sau mesaj că utilizatorul nu este în niciun joc
          content:
            application/json:
              schema:
                type: object
                properties:
                  joc:
                    $ref: '#/components/schemas/Joc'
                  gameCode:
                    type: integer
                    example: 1123
        '500':
          description: Eroare la preluarea informațiilor

  # Rută pentru pornirea jocului
  /api/jocuri/startJoc/{idJoc}:
    post:
      tags:
        - joc
      summary: Pornește un joc
      description: Schimbă starea jocului la START și inițializează cronometrul
      operationId: startGame
      parameters:
        - name: idJoc
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Jocul a fost pornit cu succes
        '400':
          description: Nu sunt suficienți jucători sau utilizatorul nu este în acest joc
        '404':
          description: Jocul nu a fost găsit

  # Rută pentru obținerea timpului de joc
  /api/jocuri/{idJoc}/timp:
    get:
      tags:
        - joc
      summary: Obține timpul jocului
      description: Returnează timpul scurs de la începerea jocului
      operationId: getGameTimer
      parameters:
        - name: idJoc
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Timp joc obținut cu succes
          content:
            application/json:
              schema:
                type: integer
                format: int64
                example: 120
        '404':
          description: Jocul nu a fost găsit
components:
  schemas:
...
    Joc:
      type: object
      properties:
        idJoc:
          type: integer
          format: int64
          example: 10
        nrJucatori:
          type: integer
          format: int64
          example: 10
        jucatori:
          type: string
          example: ralucapop;adrianalivadariu
        statusJoc:
          type: string
          example: WATING
...
    Joc-creareJoc:
      required:
        - username
      type: object
      properties:
        numarJucatori:
          type: integer
          format: int64
          example: 3
        username:
          type: string
          example: ralucapop
...
    Joc-returnareJucatoriJoc:
      type: object
      properties:
        jucatori:
          type: string
          example: ralucapop;adrianalivadariu
...
    CarduriActiveByUsername:
      type: object
      properties:
        descriereCard:
          type: string
          example: "Mergi la inchisoare"
        username:
          type: string
          example: popraluca
        cardType:
          type: string
          enum: [sansa, empire]
          example: "sansa"
        idJoc:
          type: integer
          example: 2
...
    Panou:
      type: object
      properties:
        idPanou:
          type: integer
          format: int64
          example: 10
        nume:
          type: string
          example: "McDonalds"
        pret:
          type: integer
          format: int64
          example: 100
        status:
          type: string
          enum: [activ, inactiv, general]
          example: "activ"
        idTurn:
          type: integer
          format: int64
          example: 10
...
    Turn:
      type: object
      properties:
        idTurn:
          type: integer
          format: int64
          example: 10        
        username:
          type: string
          example: "popraluca"
        idJoc:
          type: integer
          format: int64
          example: 10
        valoareTurn:
          type: integer
          format: int64
          example: 100
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header